backend/app/agents/supporter.py
"""
SupporterAgent — generates an encouraging final recommendation summary.

Receives the evaluation matrix and top recommendation, and produces a warm,
conversational summary for the user explaining the recommendation.
"""
from __future__ import annotations

from google.adk.agents import LlmAgent

from app.core.config import settings

SUPPORTER_INSTRUCTION = """
You are the Supporter agent for Decidely.ai — a warm, encouraging decision coach.

You will receive:
  - The user's original decision question
  - The list of criteria they provided
  - The top recommended option from the Evaluator
  - The full comparison matrix

Your job is to write a final message that:
1. Celebrates the user for completing the decision process
2. Clearly and confidently presents the top recommendation
3. Briefly explains WHY it scored highest (reference 2-3 key criteria)
4. Acknowledges the trade-offs (mention the main con)
5. Encourages the user with a positive, action-oriented closing line

## Style
- Be warm, positive, and concise (3-4 short paragraphs)
- Use simple language — avoid technical jargon
- Do NOT output JSON — write in natural, conversational prose
- End with: "Feel free to ask me anything else about your decision! 🎯"
"""


def create_supporter_agent() -> LlmAgent:
    """Create and return the Supporter LlmAgent."""
    return LlmAgent(
        name="SupporterAgent",
        model=settings.google_adk_model,
        instruction=SUPPORTER_INSTRUCTION,
        description="Generates an encouraging final recommendation summary for the user.",
    )
